refactoring notifications out of MapActivity and supporting a smaller Nexus 4 screen
